Remove unused EmbeddedWebserverBaseURL
It was only needed for a weird iptables workaround to run dev indico on
port 80 or 443.  Nobody is using it anymore since there are cleaner ways
such as running nginx as a frontend for the dev server or using linux
capabilities to allow Python to bind to a privileged port.